requirements.txt
snowflake-connector-python
pandas
python-dotenv
pyarrow
matplotlib
seaborn
scikit-learn
jupyter
